% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example-methods.R
\name{longForm,Example-method}
\alias{longForm,Example-method}
\alias{Example-methods}
\alias{isEmpty,Example-method}
\title{Create a long form method for Example class}
\usage{
\S4method{longForm}{Example}(object, ...)

\S4method{isEmpty}{Example}(x)
}
\arguments{
\item{object, x}{An object of class \code{Example}}

\item{...}{Additional arguments}
}
\value{
A long form of the object
}
\description{
Create a long form method for Example class
}
\examples{
ex <- Example(list(a = 1, b = 2, c = 3))
longForm(ex)
isEmpty(
    Example(x = list())
)

}
