% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example-methods.R
\name{longForm,Example-method}
\alias{longForm,Example-method}
\title{Create a long form method for Example class}
\usage{
\S4method{longForm}{Example}(object, ...)
}
\arguments{
\item{object}{An object of class \code{Example}}

\item{...}{Additional arguments}
}
\value{
A long form of the object
}
\description{
Create a long form method for Example class
}
