% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example-class.R
\docType{class}
\name{Example-class}
\alias{Example-class}
\alias{.Example}
\alias{Example}
\alias{coerce,Example,DataFrame-method}
\title{A class to represent S4 examples}
\usage{
Example(x = list())
}
\arguments{
\item{x}{Either a 'list' or a 'SimpleList' object (defaults to an empty
\code{list})}
}
\description{
This class is a simple example of an S4 class.
It inherits from the \link[S4Vectors:SimpleList-class]{SimpleList} class.
}
\section{Coercion}{

In the code example below, 'from' is an object of class \code{Example}.

\code{as(from, "DataFrame")}: Creates a \code{DataFrame} object from an \code{Example}
class. Instances of the \code{Example} class have identical lengths for all
elements.
}

\examples{
showClass("Example")

library(S4Vectors)

Example()
Example(list(a = 1, b = 2))
Example(SimpleList(a = 1, b = 2))
as(
    Example(x = list()), "DataFrame"
)
}
